+++ conf/script/src/build_system/cmd/setup/create_build_subdirs.py
